Rename Manager to BztManager and make it final
This was done in one step because G_DECLARE_FINAL_TYPE *requires* a
namespace. 'bzt' doesn't have to stick, but I figured it wasn't bad:
BlueZ Tools.

This also removes the public struct (was that even supposed to be
public?) and the public class (since we never derive it).

Lastly, several comments were removed, since they just appeared like
boilerplate. In addition, G_{BEGIN,END}_DECLS was used instead of
manual 'extern "C"'s (nevermind that we'll never build with a C++
compiler), and the indentation was fixed.