Migrate BztManager to use GInitable to initialize
GInitable is built for exactly this use-case: something needs to be
done during initialization, but it could fail. This allows the error
to bubble up into main(), where the error possibility is more
explicit. (It also allows a GCancellable, which is wired up in
BztManager here but is unused right now.)