Use macros to declare now-prefixed generated objects
The prefixes make them more obvious and hide implementation
details. It also makes them more consistent with manually-written ones
using macros, and clarifies who made e.g. Adapter.

Part of the reason I moved to make all the objects generated was to
reduce the size of diffs like this one: this would be a *massive*
commit if all of the generated files had to change.