Use GDBusProxy subclasses with GDBusObjectManagerClient
This is about three commits put together, since it turns out they were
all (somewhat) interdependent:

** Convert BztAdapter and BztDevice to GDBus subclasses.
This means that the object path can be obtained directly, removing the
need to do that in the specific classes. In addition, it allows...

** Use GDBusObjectManagerClient in BztManager.
This removes much of the ugly GVariant code that interacted with
org.freedesktop.dbus.ObjectManager, replacing it with simple
linked-list traversal.

** Return BztAdapter and BztDevices directly from bzt_manager_*
   instead of object names.
Since there's no easy way to get the Bzt* from an object path, it
becomes more convenient to pass the object. I'd like to refactor this
in future.

There are also a handful of other changes.